requirements.txt
requests>=2.32.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.15.0
python-dotenv>=1.0.0

